+++ packages/perseus/src/widgets/numeric-input/numeric-input.tsx
@@ -1,419 +1,180 @@
-import {KhanMath} from "@khanacademy/kmath";
-import {linterContextDefault} from "@khanacademy/perseus-linter";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
+import {
+    forwardRef,
+    useContext,
+    useImperativeHandle,
+    useRef,
+    useState,
+} from "react";
 import _ from "underscore";
 
 import {PerseusI18nContext} from "../../components/i18n-context";
 import InputWithExamples from "../../components/input-with-examples";
 import SimpleKeypadInput from "../../components/simple-keypad-input";
-import {ApiOptions} from "../../perseus-api";
-import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
 
-import scoreNumericInput from "./score-numeric-input";
+import {
+    NumericExampleStrings,
+    type NumericInputProps,
+} from "./numeric-input.class";
 
 import type {PerseusStrings} from "../../strings";
-import type {FocusPath, Widget, WidgetExports, WidgetProps} from "../../types";
-import type {
-    PerseusNumericInputRubric,
-    PerseusNumericInputUserInput,
-} from "../../validation.types";
-import type {NumericInputPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
-import type {
-    PerseusNumericInputWidgetOptions,
-    PerseusNumericInputAnswerForm,
-} from "@khanacademy/perseus-core";
-import type {PropsFor} from "@khanacademy/wonder-blocks-core";
+import type {PerseusNumericInputAnswerForm} from "@khanacademy/perseus-core";
 
-const formExamples: {
-    [key: string]: (
-        arg1: PerseusNumericInputAnswerForm,
-        strings: PerseusStrings,
-    ) => string;
-} = {
-    integer: (form, strings: PerseusStrings) => strings.integerExample,
-    proper: (form, strings: PerseusStrings) =>
-        form.simplify === "optional"
-            ? strings.properExample
-            : strings.simplifiedProperExample,
-    improper: (form, strings: PerseusStrings) =>
-        form.simplify === "optional"
-            ? strings.improperExample
-            : strings.simplifiedImproperExample,
-    mixed: (form, strings: PerseusStrings) => strings.mixedExample,
-    decimal: (form, strings: PerseusStrings) => strings.decimalExample,
-    pi: (form, strings: PerseusStrings) => strings.piExample,
-};
+type InputRefType = SimpleKeypadInput | InputWithExamples | null;
 
-type ExternalProps = WidgetProps<
-    PerseusNumericInputWidgetOptions,
-    PerseusNumericInputRubric
->;
+/**
+ * The NumericInputComponent is a child component of the NumericInput class
+ * component. It is responsible for rendering the UI elements of the Numeric
+ * Input widget.
+ */
+export const NumericInputComponent = forwardRef(
+    (props: NumericInputProps, ref) => {
+        const context = useContext(PerseusI18nContext);
+        const inputRef = useRef<InputRefType>(null);
+        const [isFocused, setIsFocused] = useState<boolean>(false);
 
-type Props = ExternalProps & {
-    size: NonNullable<ExternalProps["size"]>;
-    rightAlign: NonNullable<ExternalProps["rightAlign"]>;
-    apiOptions: NonNullable<ExternalProps["apiOptions"]>;
-    coefficient: NonNullable<ExternalProps["coefficient"]>;
-    answerForms: NonNullable<ExternalProps["answerForms"]>;
-    labelText: string;
-    linterContext: NonNullable<ExternalProps["linterContext"]>;
-    currentValue: string;
-};
-
-type DefaultProps = {
-    currentValue: Props["currentValue"];
-    size: Props["size"];
-    rightAlign: Props["rightAlign"];
-    apiOptions: Props["apiOptions"];
-    coefficient: Props["coefficient"];
-    answerForms: Props["answerForms"];
-    labelText: Props["labelText"];
-    linterContext: Props["linterContext"];
-};
-
-// Assert that the PerseusNumericInputWidgetOptions parsed from JSON can be passed
-// as props to this component. This ensures that the PerseusMatrixWidgetOptions
-// stays in sync with the prop types. The PropsFor<Component> type takes
-// defaultProps into account, which is important because
-// PerseusNumericInputWidgetOptions has optional fields which receive defaults
-// via defaultProps.
-0 as any as WidgetProps<
-    PerseusNumericInputWidgetOptions,
-    PerseusNumericInputRubric
-> satisfies PropsFor<typeof NumericInput>;
-
-type State = {
-    // keeps track of the other set of values when switching
-    // between 0 and finite solutions
-    previousValues: ReadonlyArray<string>;
-    isFocused: boolean;
-};
-
-export class NumericInput
-    extends React.Component<Props, State>
-    implements Widget
-{
-    static contextType = PerseusI18nContext;
-    declare context: React.ContextType<typeof PerseusI18nContext>;
-
-    inputRef: SimpleKeypadInput | InputWithExamples | null | undefined;
-
-    static defaultProps: DefaultProps = {
-        currentValue: "",
-        size: "normal",
-        rightAlign: false,
-        apiOptions: ApiOptions.defaults,
-        coefficient: false,
-        answerForms: [],
-        labelText: "",
-        linterContext: linterContextDefault,
-    };
-
-    static getUserInputFromProps(props: Props): PerseusNumericInputUserInput {
-        return {
-            currentValue: props.currentValue,
-        };
-    }
-
-    state: State = {
-        // keeps track of the other set of values when switching
-        // between 0 and finite solutions
-        previousValues: [""],
-        isFocused: false,
-    };
-
-    /**
-     * Generates a string that demonstrates how to input the various supported
-     * answer forms.
-     */
-    examples(): ReadonlyArray<string> {
-        // if the set of specified forms are empty, allow all forms
-        const forms =
-            this.props.answerForms?.length !== 0
-                ? this.props.answerForms
-                : Object.keys(formExamples).map((name) => {
-                      return {
-                          name: name,
-                          simplify: "required",
-                      } as PerseusNumericInputAnswerForm;
-                  });
-
-        let examples = _.map(forms, (form) => {
-            return formExamples[form.name](form, this.context.strings);
-        });
-        // Ensure no duplicate tooltip text from simplified and unsimplified
-        // versions of the same format
-        examples = _.uniq(examples);
-
-        return [this.context.strings.yourAnswer].concat(examples);
-    }
-
-    shouldShowExamples: () => boolean = () => {
-        const noFormsAccepted = this.props.answerForms?.length === 0;
-        // To check if all answer forms are accepted, we must first
-        // find the *names* of all accepted forms, and see if they are
-        // all present, ignoring duplicates
-        const answerFormNames: ReadonlyArray<string> = _.uniq(
-            this.props.answerForms?.map((form) => form.name),
-        );
-        const allFormsAccepted =
-            answerFormNames.length >= Object.keys(formExamples).length;
-        return !noFormsAccepted && !allFormsAccepted;
-    };
-
-    focus: () => boolean = () => {
-        this.inputRef?.focus();
-        return true;
-    };
-
-    focusInputPath: () => void = () => {
-        this.inputRef?.focus();
-    };
-
-    blurInputPath: () => void = () => {
-        this.inputRef?.blur();
-    };
-
-    getInputPaths: () => ReadonlyArray<ReadonlyArray<string>> = () => {
-        // The widget itself is an input, so we return a single empty list to
-        // indicate this.
-        /* c8 ignore next */
-        return [[]];
-    };
-
-    setInputValue: (
-        arg1: FocusPath,
-        arg2: string,
-        arg3?: () => unknown | null | undefined,
-    ) => void = (path, newValue, cb) => {
-        /* c8 ignore next */
-        this.props.onChange(
-            {
-                currentValue: newValue,
+        // Pass the focus and blur methods to the Numeric Input Class component
+        useImperativeHandle(ref, () => ({
+            focus: () => {
+                if (inputRef.current) {
+                    inputRef.current.focus();
+                    setIsFocused(true);
+                }
             },
-            cb,
-        );
-    };
+            blur: () => {
+                if (inputRef.current) {
+                    inputRef.current.blur();
+                    setIsFocused(false);
+                }
+            },
+        }));
 
-    getUserInput(): PerseusNumericInputUserInput {
-        return NumericInput.getUserInputFromProps(this.props);
-    }
+        const handleChange = (
+            newValue: string,
+            cb?: () => unknown | null | undefined,
+        ): void => {
+            props.onChange({currentValue: newValue}, cb);
+            props.trackInteraction();
+        };
 
-    getPromptJSON(): NumericInputPromptJSON {
-        return _getPromptJSON(this.props, this.getUserInput());
-    }
+        const handleFocus = (): void => {
+            props.onFocus([]);
+            setIsFocused(true);
+        };
 
-    handleChange: (
-        arg1: string,
-        arg2?: () => unknown | null | undefined,
-    ) => void = (newValue, cb) => {
-        this.props.onChange({currentValue: newValue}, cb);
-        this.props.trackInteraction();
-    };
+        const handleBlur = (): void => {
+            props.onBlur([]);
+            setIsFocused(false);
+        };
 
-    _handleFocus: () => void = () => {
-        this.props.onFocus([]);
-        this.setState((currentState) => {
-            return {...currentState, isFocused: true};
-        });
-    };
-
-    _handleBlur: () => void = () => {
-        this.props.onBlur([]);
-        this.setState((currentState) => {
-            return {...currentState, isFocused: false};
-        });
-    };
-
-    render(): React.ReactNode {
-        let labelText = this.props.labelText;
-        if (labelText == null || labelText === "") {
-            labelText = this.context.strings.yourAnswerLabel;
-        }
-
-        // To right align a custom keypad we need to wrap it.
+        /**
+         * Add right alignment to the keypad input if necessary.
+         */
         const maybeRightAlignKeypadInput = (
             keypadInput: React.ReactElement<
                 React.ComponentProps<typeof SimpleKeypadInput>
             >,
         ) => {
-            return this.props.rightAlign ? (
+            return props.rightAlign ? (
                 <div className="perseus-input-right-align">{keypadInput}</div>
             ) : (
                 keypadInput
             );
         };
 
-        if (this.props.apiOptions.customKeypad) {
-            // TODO(charlie): Support "Review Mode".
-            return maybeRightAlignKeypadInput(
-                <SimpleKeypadInput
-                    ref={(ref) => (this.inputRef = ref)}
-                    value={this.props.currentValue}
-                    keypadElement={this.props.keypadElement}
-                    onChange={this.handleChange}
-                    onFocus={this._handleFocus}
-                    onBlur={this._handleBlur}
-                />,
-            );
+        // If the labelText is not provided by the Content Creators, use the default label text
+        let labelText = props.labelText;
+        if (labelText == null || labelText === "") {
+            labelText = context.strings.yourAnswerLabel;
         }
 
-        // Note: This is _very_ similar to what `input-number.jsx` does. If
-        // you modify this, double-check if you also need to modify that
-        // component.
+        // Styles for the InputWithExamples
         const styles = StyleSheet.create({
             input: {
                 borderRadius: "3px",
-                borderWidth: this.state.isFocused ? "2px" : "1px",
+                borderWidth: isFocused ? "2px" : "1px",
                 display: "inline-block",
                 fontFamily: `Symbola, "Times New Roman", serif`,
                 fontSize: "18px",
                 height: "32px",
                 lineHeight: "18px",
-                padding: this.state.isFocused ? "4px" : "4px 5px", // account for added focus border thickness
-                textAlign: this.props.rightAlign ? "right" : "left",
-                width: this.props.size === "small" ? 40 : 80,
+                padding: isFocused ? "4px" : "4px 5px",
+                textAlign: props.rightAlign ? "right" : "left",
+                width: props.size === "small" ? 40 : 80,
             },
         });
 
+        // (mobile) If the custom keypad is enabled, use the SimpleKeypadInput component
+        if (props.apiOptions.customKeypad) {
+            return maybeRightAlignKeypadInput(
+                <SimpleKeypadInput
+                    ref={inputRef as React.RefObject<SimpleKeypadInput>}
+                    value={props.currentValue}
+                    keypadElement={props.keypadElement}
+                    onChange={handleChange}
+                    onFocus={handleFocus}
+                    onBlur={handleBlur}
+                />,
+            );
+        }
+        // (desktop) Otherwise, use the InputWithExamples component
         return (
             <InputWithExamples
-                ref={(ref) => (this.inputRef = ref)}
-                value={this.props.currentValue}
-                onChange={this.handleChange}
+                ref={inputRef as React.RefObject<InputWithExamples>}
+                value={props.currentValue}
+                onChange={handleChange}
                 labelText={labelText}
-                examples={this.examples()}
-                shouldShowExamples={this.shouldShowExamples()}
-                onFocus={this._handleFocus}
-                onBlur={this._handleBlur}
-                id={this.props.widgetId}
-                disabled={this.props.apiOptions.readOnly}
+                examples={generateExamples(props.answerForms, context.strings)}
+                shouldShowExamples={shouldShowExamples(props.answerForms)}
+                onFocus={handleFocus}
+                onBlur={handleBlur}
+                id={props.widgetId}
+                disabled={props.apiOptions.readOnly}
                 style={styles.input}
             />
         );
-    }
-}
-
-// TODO(thomas): Currently we receive a list of lists of acceptable answer types
-// and union them down into a single set. It's worth considering whether it
-// wouldn't make more sense to have a single set of acceptable answer types for
-// a given *problem* rather than for each possible [correct/wrong] *answer*.
-// When should two answers to a problem take different answer types?
-// See D27790 for more discussion.
-export const unionAnswerForms: (
-    arg1: ReadonlyArray<ReadonlyArray<PerseusNumericInputAnswerForm>>,
-) => ReadonlyArray<PerseusNumericInputAnswerForm> = function (answerFormsList) {
-    // Takes a list of lists of answer forms, and returns a list of the forms
-    // in each of these lists in the same order that they're listed in the
-    // `formExamples` forms from above.
-
-    // uniqueBy takes a list of elements and a function which compares whether
-    // two elements are equal, and returns a list of unique elements. This is
-    // just a helper function here, but works generally.
-    const uniqueBy = function (list, iteratee: any) {
-        // @ts-expect-error - TS2347 - Untyped function calls may not accept type arguments.
-        return list.reduce<Array<any>>((uniqueList, element) => {
-            // For each element, decide whether it's already in the list of
-            // unique items.
-            const inList = _.find(uniqueList, iteratee.bind(null, element));
-            if (inList) {
-                return uniqueList;
-            }
-            return uniqueList.concat([element]);
-        }, []);
-    };
-
-    // Pull out all of the forms from the different lists.
-    const allForms = answerFormsList.flat();
-    // Pull out the unique forms using uniqueBy.
-    const uniqueForms = uniqueBy(allForms, _.isEqual);
-    // Sort them by the order they appear in the `formExamples` list.
-    const formExampleKeys = Object.keys(formExamples);
-    return _.sortBy(uniqueForms, (form) => {
-        return formExampleKeys.indexOf(form.name);
-    });
-};
-
-type RenderProps = {
-    answerForms: ReadonlyArray<{
-        simplify: "required" | "correct" | "enforced" | null | undefined;
-        name: "integer" | "decimal" | "proper" | "improper" | "mixed" | "pi";
-    }>;
-    labelText: string;
-    size: "normal" | "small";
-    coefficient: boolean;
-    rightAlign?: boolean;
-    static: boolean;
-};
-
-const propsTransform = function (
-    widgetOptions: PerseusNumericInputWidgetOptions,
-): RenderProps {
-    const rendererProps = _.extend(_.omit(widgetOptions, "answers"), {
-        answerForms: unionAnswerForms(
-            // Pull out the name of each form and whether that form has
-            // required simplification.
-            widgetOptions.answers.map((answer) => {
-                // @ts-expect-error - TS2345 - Argument of type 'readonly MathFormat[] | undefined' is not assignable to parameter of type 'Collection<any>'.
-                return _.map(answer.answerForms, (form) => {
-                    return {
-                        simplify: answer.simplify,
-                        name: form,
-                    };
-                });
-            }),
-        ),
-    });
-
-    return rendererProps;
-};
-
-export default {
-    name: "numeric-input",
-    displayName: "Numeric input",
-    defaultAlignment: "inline-block",
-    accessible: true,
-    widget: NumericInput,
-    transform: propsTransform,
-    isLintable: true,
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type 'UserInput' is not assignable to type 'PerseusNumericInputUserInput'.
-    scorer: scoreNumericInput,
-
-    // TODO(LEMS-2656): remove TS suppression
-    // @ts-expect-error: Type 'Rubric' is not assignable to type 'PerseusNumericInputRubric'
-    getOneCorrectAnswerFromRubric(
-        rubric: PerseusNumericInputRubric,
-    ): string | null | undefined {
-        const correctAnswers = rubric.answers.filter(
-            (answer) => answer.status === "correct",
-        );
-        const answerStrings = correctAnswers.map((answer) => {
-            // Figure out how this answer is supposed to be
-            // displayed
-            let format = "decimal";
-            if (answer.answerForms && answer.answerForms[0]) {
-                // NOTE(johnsullivan): This isn't exactly ideal, but
-                // it does behave well for all the currently known
-                // problems. See D14742 for some discussion on
-                // alternate strategies.
-                format = answer.answerForms[0];
-            }
-
-            // @ts-expect-error - TS2345 - Argument of type 'string' is not assignable to parameter of type 'MathFormat | undefined'.
-            let answerString = KhanMath.toNumericString(answer.value, format);
-            if (answer.maxError) {
-                answerString +=
-                    " \u00B1 " +
-                    // @ts-expect-error - TS2345 - Argument of type 'string' is not assignable to parameter of type 'MathFormat | undefined'.
-                    KhanMath.toNumericString(answer.maxError, format);
-            }
-            return answerString;
-        });
-        if (answerStrings.length === 0) {
-            return;
-        }
-        return answerStrings[0];
     },
-} satisfies WidgetExports<typeof NumericInput>;
+);
+
+/**
+ * Generates a string that demonstrates how to input the various supported
+ * answer forms. These strings are shown as examples to the user in a tooltip.
+ */
+const generateExamples = (
+    answerForms: readonly PerseusNumericInputAnswerForm[],
+    strings: PerseusStrings,
+): ReadonlyArray<string> => {
+    const forms =
+        answerForms?.length !== 0
+            ? answerForms
+            : Object.keys(NumericExampleStrings).map((name) => {
+                  return {
+                      name: name,
+                      simplify: "required",
+                  } as PerseusNumericInputAnswerForm;
+              });
+
+    let examples = _.map(forms, (form) => {
+        return NumericExampleStrings[form.name](form, strings);
+    });
+    examples = _.uniq(examples);
+
+    return [strings.yourAnswer].concat(examples);
+};
+
+/**
+ * Determines whether to show examples of how to input
+ * the various supported answer forms.
+ */
+const shouldShowExamples = (
+    answerForms: readonly PerseusNumericInputAnswerForm[],
+): boolean => {
+    const noFormsAccepted = answerForms?.length === 0;
+    const answerFormNames: ReadonlyArray<string> = _.uniq(
+        answerForms?.map((form) => form.name),
+    );
+    const allFormsAccepted =
+        answerFormNames.length >= Object.keys(NumericExampleStrings).length;
+    return !noFormsAccepted && !allFormsAccepted;
+};
